+++ resources/org.nodec.hello-nodec-game/prefabs/bullet.entity
@@ -3,7 +3,7 @@
     "components": [
       {
         "polymorphic_id": 2147483649,
-        "polymorphic_name": "components::Bullet",
+        "polymorphic_name": "hello_nodec_game::components::Bullet",
         "ptr_wrapper": {
           "valid": 1,
           "data": {}
