+++ resources/org.nodec.hello-nodec-game/prefabs/main-scene.entity
@@ -1635,7 +1635,7 @@
             "components": [
               {
                 "polymorphic_id": 2147483658,
-                "polymorphic_name": "components::PlayerControl",
+                "polymorphic_name": "hello_nodec_game::components::PlayerControl",
                 "ptr_wrapper": {
                   "valid": 1,
                   "data": {
