+++ src/components/bullet.hpp
@@ -1,12 +1,9 @@
-#ifndef COMPONENTS__BULLET_HPP_
-#define COMPONENTS__BULLET_HPP_
+#ifndef HELLO_NODEC_GAME__COMPONENTS__BULLET_HPP_
+#define HELLO_NODEC_GAME__COMPONENTS__BULLET_HPP_
 
 #include <nodec_scene_serialization/serializable_component.hpp>
 
-#ifdef EDITOR_MODE
-#    include <nodec_scene_editor/scene_editor.hpp>
-#endif
-
+namespace hello_nodec_game {
 namespace components {
 
 class Bullet : public nodec_scene_serialization::BaseSerializableComponent {
@@ -19,20 +16,10 @@ public:
     template<class Archive>
     void serialize(Archive &archive) {
     }
-
-#ifdef EDITOR_MODE
-public:
-    static void setup_editor(nodec_scene_editor::SceneEditor &editor) {
-        editor.inspector_component_registry().register_component<Bullet>(
-            "Bullet",
-            [](Bullet &bullet) {
-
-            });
-    }
-#endif
 };
 
 } // namespace components
+} // namespace hello_nodec_game
 
-NODEC_SCENE_REGISTER_SERIALIZABLE_COMPONENT(components::Bullet)
+NODEC_SCENE_REGISTER_SERIALIZABLE_COMPONENT(hello_nodec_game::components::Bullet)
 #endif
